Wait for jobs to complete before notifying
Previously, we would send notifications effectively as soon as the
post-exec script told us that the job was done, plus a short delay to
let LSF sort its life out. This meant we'd sometimes send a message
saying "your job will be done in a moment" (because we still observed a
RUN state) rather than describing whether the job succeeded (DONE) or
failed (EXIT). Now we should always be able to include that information.

This approach also lets us properly handle job arrays. Previously, there
was a chance that a notification for a job array might go missing: if
two jobs in the same array were to finish simultaneously, we'd only let
one of the handlers test for a DONE/EXIT state, to avoid sending
multiple notifications for a single job array (and because, at the time,
there was no batching of bjobs invocations, so 100 handlers asking about
the same array once each would have used up all bjobs quota for a
quarter of an hour). But if that handler happened to see a state where
one member of the array was still RUNning, then it would skip the
notification, and assume that another handler would fire later.

Now, all handlers for a job array will wait for their own job to finish,
then wait for the rest of the array's handlers to do so, and finally a
single handler will be elected to send the notification. This approach
is not particularly robust in the case of a single job whose post-exec
script fails, but it does avoid polling bjobs forever if part of an
array goes missing.